refactor: move validateEventGroupSelectors to RequisitionsValidator
Move validateEventGroupSelectors and InconsistentEventGroupSelectorsException
from RequisitionGrouper to RequisitionsValidator, co-locating all
requisition validation logic in a single class.
