+++ src/main/kotlin/org/wfanet/measurement/edpaggregator/requisitionfetcher/RequisitionsValidator.kt
@@ -28,6 +28,7 @@ import org.wfanet.measurement.api.v2alpha.unpack
 import org.wfanet.measurement.common.crypto.PrivateKeyHandle
 import org.wfanet.measurement.consent.client.dataprovider.decryptRequisitionSpec
 import org.wfanet.measurement.edpaggregator.v1alpha.GroupedRequisitions
+import org.wfanet.measurement.edpaggregator.v1alpha.GroupedRequisitions.EventGroupDetails
 
 class InvalidRequisitionException(
   val requisitions: List<Requisition>,
@@ -129,7 +130,35 @@ class RequisitionsValidator(private val privateEncryptionKey: PrivateKeyHandle) 
     }
   }
 
+  /**
+   * Validates that event group selectors are consistent across all event groups.
+   *
+   * If any event group has an `entity_key`, all must. If none have an `entity_key`, all must have a
+   * non-empty `event_group_reference_id`.
+   */
+  fun validateEventGroupSelectors(eventGroupMap: Map<String, EventGroupDetails>) {
+    require(eventGroupMap.isNotEmpty()) { "eventGroupMap must not be empty" }
+
+    val hasEntityKey = eventGroupMap.values.any { it.hasEntityKey() }
+    if (hasEntityKey) {
+      if (!eventGroupMap.values.all { it.hasEntityKey() }) {
+        throw InconsistentEventGroupSelectorsException(
+          "Inconsistent selectors: if any event group has entity_key, all must"
+        )
+      }
+    } else {
+      if (!eventGroupMap.values.all { it.eventGroupReferenceId.isNotEmpty() }) {
+        throw InconsistentEventGroupSelectorsException(
+          "All event groups must have event_group_reference_id when entity_key is not present"
+        )
+      }
+    }
+  }
+
   companion object {
     private val logger: Logger = Logger.getLogger(this::class.java.name)
   }
 }
+
+/** Thrown when event group selectors (entity_key vs event_group_reference_id) are inconsistent. */
+class InconsistentEventGroupSelectorsException(message: String) : Exception(message)
